requirements.txt
azure-core==1.35.0
azure-data-tables==12.7.0
azure-identity==1.23.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
ecdsa==0.19.1
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
propcache==0.3.2
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-http-client==3.3.7
requests==2.32.4
sendgrid==6.12.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
